backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-community==0.3.27
langchain-text-splitters==0.3.9
faiss-cpu==1.11.0.post1
sentence-transformers==5.1.0
groq==0.31.0
gitpython==3.1.45
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
